+++ packages/evm/source/index.ts
@@ -1,13 +1,13 @@
 import { Identifiers } from "@mainsail/contracts";
 import { Providers } from "@mainsail/kernel";
 
-import { Instance } from "./instance";
+import { EvmInstance } from "./instances/evm";
 
 export * as Bindings from "./generated/bindings";
 
 export class ServiceProvider extends Providers.ServiceProvider {
 	public async register(): Promise<void> {
-		this.app.bind(Identifiers.Evm.Instance).to(Instance).inSingletonScope();
+		this.app.bind(Identifiers.Evm.Instance).to(EvmInstance).inSingletonScope();
 	}
 
 	public async boot(): Promise<void> {}
