+++ packages/evm/source/instances/evm.ts
@@ -1,10 +1,10 @@
 import { injectable } from "@mainsail/container";
 import { Contracts } from "@mainsail/contracts";
 
-import { Evm, JsTransactionContext, JsTransactionResult } from "./generated/bindings";
+import { Evm, JsTransactionContext, JsTransactionResult } from "../generated/bindings";
 
 @injectable()
-export class Instance implements Contracts.Evm.Instance {
+export class EvmInstance implements Contracts.Evm.Instance {
 	private readonly evm: Evm = new Evm();
 
 	public async transact(txContext: JsTransactionContext): Promise<JsTransactionResult> {
